requirements.txt
discord.py[voice]
yt_dlp
spotipy
python-dotenv